refactor(storage): replace AsyncRead with PieceContentStream in clients
- Introduce `PieceContentStream` type alias (`BoxStream` of `Bytes`) in
  the storage client module.
- Replace `impl AsyncRead` returns with `PieceContentStream` in both
  QUIC and TCP clients.
- QUIC client wraps `RecvStream` via `read_chunk` for zero-copy chunks;
  TCP client uses `ReaderStream` with configurable buffer size.
- Move `FDCache`, `BufferPool`, and related types to dedicated utility
  modules, cleaning up `content.rs`.
- Add `write_piece_from_stream` variants everywhere, dropping the old
  `AsyncRead`-based write paths.
