+++ dragonfly-client-storage/src/client/tcp.rs
@@ -17,11 +17,13 @@
 use bytes::{Bytes, BytesMut};
 use dragonfly_client_config::dfdaemon::Config;
 use dragonfly_client_core::{Error as ClientError, Result as ClientResult};
+use futures::StreamExt;
 use socket2::{SockRef, TcpKeepalive};
 use std::sync::Arc;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncWriteExt};
+use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::net::tcp::{OwnedReadHalf, OwnedWriteHalf};
 use tokio::time;
+use tokio_util::io::ReaderStream;
 use tracing::{debug, error, instrument, Span};
 use vortex_protocol::{
     tlv::{
@@ -50,6 +52,12 @@ impl TCPClient {
         Self { config, addr }
     }
 
+    /// Converts the reader holding the piece content into a stream of bytes
+    /// chunks, so the storage writes the chunks without copying them again.
+    fn content_stream(&self, reader: OwnedReadHalf) -> super::PieceContentStream {
+        ReaderStream::with_capacity(reader, self.config.storage.write_buffer_size).boxed()
+    }
+
     /// Downloads a piece from the server using the vortex protocol.
     ///
     /// This is the main entry point for downloading a piece. It applies
@@ -59,7 +67,7 @@ impl TCPClient {
         &self,
         number: u32,
         task_id: &str,
-    ) -> ClientResult<(impl AsyncRead, u64, String)> {
+    ) -> ClientResult<(super::PieceContentStream, u64, String)> {
         Span::current().record("parent_addr", self.addr.as_str());
 
         time::timeout(
@@ -83,7 +91,7 @@ impl TCPClient {
         &self,
         number: u32,
         task_id: &str,
-    ) -> ClientResult<(impl AsyncRead, u64, String)> {
+    ) -> ClientResult<(super::PieceContentStream, u64, String)> {
         let request: Bytes = Vortex::DownloadPiece(
             Header::new_download_piece(),
             DownloadPiece::new(task_id.to_string(), number),
@@ -100,7 +108,11 @@ impl TCPClient {
                 debug!("received piece content: {:?}", piece_content.metadata());
 
                 let metadata = piece_content.metadata();
-                Ok((reader, metadata.offset, metadata.digest))
+                Ok((
+                    self.content_stream(reader),
+                    metadata.offset,
+                    metadata.digest,
+                ))
             }
             Tag::Error => Err(self.read_error(&mut reader, header.length() as usize).await),
             _ => Err(ClientError::Unknown(format!(
@@ -118,7 +130,7 @@ impl TCPClient {
         &self,
         number: u32,
         task_id: &str,
-    ) -> ClientResult<(impl AsyncRead, u64, String)> {
+    ) -> ClientResult<(super::PieceContentStream, u64, String)> {
         time::timeout(
             self.config.download.piece_timeout,
             self.handle_download_persistent_piece(number, task_id),
@@ -138,7 +150,7 @@ impl TCPClient {
         &self,
         number: u32,
         task_id: &str,
-    ) -> ClientResult<(impl AsyncRead, u64, String)> {
+    ) -> ClientResult<(super::PieceContentStream, u64, String)> {
         let request: Bytes = Vortex::DownloadPersistentPiece(
             Header::new_download_persistent_piece(),
             DownloadPersistentPiece::new(task_id.to_string(), number),
@@ -158,7 +170,11 @@ impl TCPClient {
                 );
 
                 let metadata = persistent_piece_content.metadata();
-                Ok((reader, metadata.offset, metadata.digest))
+                Ok((
+                    self.content_stream(reader),
+                    metadata.offset,
+                    metadata.digest,
+                ))
             }
             Tag::Error => Err(self.read_error(&mut reader, header.length() as usize).await),
             _ => Err(ClientError::Unknown(format!(
@@ -176,7 +192,7 @@ impl TCPClient {
         &self,
         number: u32,
         task_id: &str,
-    ) -> ClientResult<(impl AsyncRead, u64, String)> {
+    ) -> ClientResult<(super::PieceContentStream, u64, String)> {
         time::timeout(
             self.config.download.piece_timeout,
             self.handle_download_persistent_cache_piece(number, task_id),
@@ -196,7 +212,7 @@ impl TCPClient {
         &self,
         number: u32,
         task_id: &str,
-    ) -> ClientResult<(impl AsyncRead, u64, String)> {
+    ) -> ClientResult<(super::PieceContentStream, u64, String)> {
         let request: Bytes = Vortex::DownloadPersistentCachePiece(
             Header::new_download_persistent_cache_piece(),
             DownloadPersistentCachePiece::new(task_id.to_string(), number),
@@ -216,7 +232,11 @@ impl TCPClient {
                 );
 
                 let metadata = persistent_cache_piece_content.metadata();
-                Ok((reader, metadata.offset, metadata.digest))
+                Ok((
+                    self.content_stream(reader),
+                    metadata.offset,
+                    metadata.digest,
+                ))
             }
             Tag::Error => Err(self.read_error(&mut reader, header.length() as usize).await),
             _ => Err(ClientError::Unknown(format!(
